Struct probability::prelude::Pert

pub struct Pert { /* fields omitted */ }

A PERT distribution.

Implementations

impl Pert

pub fn new(a: f64, b: f64, c: f64) -> Self

Create a PERT distribution with parameters a, b, and c.

It should hold that a < b < c.

pub fn a(&self) -> f64

Return the first parameter.

pub fn b(&self) -> f64

Return the second parameter.

pub fn c(&self) -> f64

Return the third parameter.

pub fn alpha(&self) -> f64

Return the first shape parameter of the corresponding Beta distribution.

pub fn beta(&self) -> f64

Return the second shape parameter of the corresponding Beta distribution.