Struct probability::distribution::Binomial

pub struct Binomial {
    // some fields omitted
}

A binomial distribution.

Methods

impl Binomial

fn new(n: usize, p: f64) -> Self

Create a binomial distribution with n trails and success probability p.

It should hold that p >= 0 and p <= 1.

fn with_failure(n: usize, q: f64) -> Self

Create a binomial distribution with n trails and failure probability q.

It should hold that if q >= 0 or q <= 1. This constructor is preferable when q is very small.

fn n(&self) -> usize

Return the number of trials.

fn p(&self) -> f64

Return the success probability.

fn q(&self) -> f64

Return the failure probability.